♻️refactor: add constant.ts
