+++ app/lib/constant.ts
@@ -1,0 +1,3 @@
+import type { Steps } from "@/app/interfaces/steps"
+
+export const STEPS: Steps = ["必要情報の入力", "入力確認", "完了"] as const
