+++ app/register/page.tsx
@@ -1,9 +1,7 @@
 import { Stepper } from "@/app/components/stepper"
-import type { Steps } from "@/app/interfaces/steps"
+import { STEPS } from "@/app/lib/constant"
 import type React from "react"
 import { ProfileForm } from "./components/profileForm"
-
-export const STEPS: Steps = ["必要情報の入力", "入力確認", "完了"] as const
 
 export default function Register(): React.JSX.Element {
   return (
