+++ app/register/success/page.tsx
@@ -1,5 +1,5 @@
 import { Stepper } from "@/app/components/stepper"
-import { STEPS } from "@/app/register/page"
+import { STEPS } from "@/app/lib/constant"
 import type React from "react"
 
 export default function Success(): React.JSX.Element {
